fix fitness

+++ src/Entities/TSPDecoder.cpp
@@ -12,14 +12,14 @@ TSPDecoder::~TSPDecoder()
 
 BRKGA::fitness_t TSPDecoder::decode(BRKGA::Chromosome &chromosome, bool /* not-used */)
 {
-	uint fitness = instance.evaluate(std::vector<double>(chromosome.begin(), chromosome.end()));
+	double fitness = instance.evaluate(std::vector<double>(chromosome.begin(), chromosome.end()));
 	instance.reset();
 	return fitness;
 }
 
 void TSPDecoder::printSolution(BRKGA::Chromosome &chromosome)
 {
-	uint fitness = instance.evaluate(std::vector<double>(chromosome.begin(), chromosome.end()));
+	double fitness = instance.evaluate(std::vector<double>(chromosome.begin(), chromosome.end()));
 	instance.printStatistics();
 	std::cout << "Best cost: " << fitness << std::endl
 			  << std::endl;
